refactor: create Frame controller
